Layout: Small refactor of layout into a tree.
Currently we just have the Document root node, and the Block node containing all the content. However I did verify that the code still works after this change so the refactor is less risky.

Currently up to the text in the book "Make sure to rename the Layout constructor call in DocumentLayout as well. As always, test your browser and make sure that after all of these refactors, everything still works."